Added CDN references for MVC's jQuery and AJAX
support. Also removed references to local jquery- ui files bundled with
the timepicker library - they seemed to be interfering with standard
jquery-ui dialog() calls, and the timepicker library seems to work just
fine with the standard ui library. (I checked both the website and the
functionality of the timepicker in the app.)