Fixed a bug with the search functionality. When a
search query was executed, result courses are added to a list, if any
linked courses exist they are Unioned to the existing list. This list is
then passed to the Search view, which defines a List collection as the
Model. However, the Union method was returning an IEnumerable rather
than a List. So any search query that had linked courses (requiring a
Union) would throw an exception on the Search view. This was fixed by
converting IEnumerable to List before passing to the view. The fix was
implemented on both the YearQuarterSubject and Search/Index views.