Changed the way course descriptions are displayed.
The order that they were assumed to be in (current description to
upcoming) was not accurate. The Course object's list of
CourseDescriptions seems to be a list of all descriptions. The code now
implements logic to determined the current, and any upcoming
descriptions. The method which contains this logic has been abstracted
into the Helpers class so that both Views (ClassDetails and Subject)
which must make use of the logic, can.