Fixed issue where it was impossible to detect
which subject a course belonged to. The relationship between course
prefixes and subjects meant that a prefix could belong to multiple
subjects. This caused problems in the SubjectInfo class, and Class
Details page. The relationship has been modified so that any course
prefix can only belong to a single subject. The code has been modified
to reflect this new relationship.