Fixed browsing to search page error
FogBugz:4554
If no search term is present, the user is redirected back to the home
page now. (There is no search page - only a search results page.) I
also fixed it so that you can manually specify a search term in the
URL without the quarter - the search will default to the currently-
active quarter.